+++ packages/waku/src/lib/middleware/handler.ts
@@ -79,7 +79,7 @@ export const handler: Middleware = (options) => {
       entriesPrd,
     ] = await Promise.all([configPromise, entriesPromise]);
     const entriesDev = devServer && (await devServer.loadEntriesDev(config));
-    const entries: { default: object } = devServer ? entriesDev! : entriesPrd;
+    const entries = devServer ? entriesDev! : entriesPrd;
     const rsdwServer = devServer
       ? await devServer.loadServerModuleRsc(SERVER_MODULE_MAP['rsdw-server'])
       : await entriesPrd.loadModule('rsdw-server');
@@ -140,30 +140,27 @@ export const handler: Middleware = (options) => {
         };
       },
     };
-    if ('unstable_handleRequest' in entries.default) {
-      const input = await getInput(config, ctx, loadServerModule);
-      if (input) {
-        const res = await (
-          entries.default.unstable_handleRequest as HandleRequest
-        )(input, utils);
-        if (res instanceof ReadableStream) {
-          ctx.res.body = res;
-        } else if (res) {
-          if (res.body) {
-            ctx.res.body = res.body;
-          }
-          if (res.status) {
-            ctx.res.status = res.status;
-          }
-          if (res.headers) {
-            Object.assign((ctx.res.headers ||= {}), res.headers);
-          }
+    const input = await getInput(config, ctx, loadServerModule);
+    if (input) {
+      const res = await entries.default.unstable_handleRequest(input, utils);
+      if (res instanceof ReadableStream) {
+        ctx.res.body = res;
+      } else if (res) {
+        if (res.body) {
+          ctx.res.body = res.body;
         }
-        if (ctx.res.body || ctx.res.status) {
-          return;
+        if (res.status) {
+          ctx.res.status = res.status;
+        }
+        if (res.headers) {
+          Object.assign((ctx.res.headers ||= {}), res.headers);
         }
       }
+      if (ctx.res.body || ctx.res.status) {
+        return;
+      }
     }
+
     await next();
   };
 };
