+++ packages/waku/src/minimal/server.ts
@@ -1,16 +1,9 @@
 import type { ReactNode } from 'react';
 
-import type { Config } from '../config.js';
 import type { PathSpec } from '../lib/utils/path.js';
-// TODO move types somewhere
-import type { HandlerReq, HandlerRes } from '../lib/middleware/types.js';
+import type { HandlerReq, HandlerRes } from '../lib/types.js';
 
 type Elements = Record<string, ReactNode>;
-
-// -----------------------------------------------------
-// new_defineEntries
-// Eventually replaces defineEntries
-// -----------------------------------------------------
 
 type HandleRequest = (
   input: (
@@ -37,7 +30,7 @@ type HandleRequest = (
   },
 ) => Promise<ReadableStream | HandlerRes | null | undefined>;
 
-export type new_BuildConfig = {
+type BuildConfig = {
   pathSpec: PathSpec;
   isStatic?: boolean | undefined;
   entries?: {
@@ -48,25 +41,13 @@ export type new_BuildConfig = {
   customCode?: string; // optional code to inject TODO hope to remove this
 }[];
 
-type new_GetBuildConfig = (utils: {
+type GetBuildConfig = (utils: {
   unstable_collectClientModules: (elements: Elements) => Promise<string[]>;
-}) => Promise<new_BuildConfig>;
+}) => Promise<BuildConfig>;
 
 export function new_defineEntries(fns: {
   unstable_handleRequest: HandleRequest;
-  unstable_getBuildConfig: new_GetBuildConfig;
+  unstable_getBuildConfig: GetBuildConfig;
 }) {
   return fns;
 }
-
-export type EntriesDev = {
-  default: ReturnType<typeof new_defineEntries>;
-};
-
-export type EntriesPrd = EntriesDev & {
-  loadConfig: () => Promise<Config>;
-  loadModule: (id: string) => Promise<unknown>;
-  dynamicHtmlPaths: [pathSpec: PathSpec, htmlHead: string][];
-  publicIndexHtml: string;
-  buildData?: Record<string, unknown>; // must be JSON serializable
-};
